requirements.txt
selene>=2.0.0rc7
pytest>=8.3.3
selenium~=4.26.1
allure-pytest==2.9.45
allure-python-commons~=2.13.5